+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Added
 
 - Prebuilt linux-riscv64 binary release.
+- Support for CRSH greeting message format 2.
 
 ## [0.5] - 2026-03-18